Move unit tests in dedicated folder #14
